requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
websockets==15.0
python-multipart==0.0.20
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
numpy==2.3.1
pyfiglet==1.0.3
requests==2.32.4
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
psycopg2-binary==2.9.10
supabase==2.11.0
aiofiles==24.1.0
